Extract DDL types to separate module
- Create crates/ast/src/ddl.rs with CreateTableStmt and ColumnDef
- Add module declaration and public exports to lib.rs
- Remove old DDL definitions from lib.rs (lines 137-150)
- All 22 AST tests pass
- No clippy warnings
- ddl.rs: 19 lines, lib.rs reduced to 549 lines

Closes #30
