+++ crates/ast/src/ddl.rs
@@ -1,0 +1,19 @@
+//! Data Definition Language (DDL) AST nodes
+
+use types::DataType;
+
+/// CREATE TABLE statement
+#[derive(Debug, Clone, PartialEq)]
+pub struct CreateTableStmt {
+    pub table_name: String,
+    pub columns: Vec<ColumnDef>,
+    // TODO: Add constraints
+}
+
+/// Column definition
+#[derive(Debug, Clone, PartialEq)]
+pub struct ColumnDef {
+    pub name: String,
+    pub data_type: DataType,
+    pub nullable: bool,
+}
